- Refactor options
