--- lib/icodi/option_handling.rb
@@ -1,47 +1,0 @@
-module IcodiCore
-  module OptionHandling
-    def method_missing(method_name, *_args, &_block)
-      respond_to?(method_name) ? options[method_name] : super
-    end
-
-    def respond_to_missing?(method_name, _include_private = false)
-      options.has_key?(method_name) ? true : super
-    end
-
-    def default_options
-      {
-        template:   :default,
-        pixels:     5,
-        density:    0.5,
-        stroke:     0.1,
-        background: '#fff',
-        color:      random_color,
-        mirror:     :x,
-        jitter:     0,
-        id:         :icodi,
-      }
-    end
-
-    # Derivative Properties
-
-    def size
-      @size ||= pixels * 10
-    end
-
-    def style
-      @style ||= { stroke: color, stroke_width: stroke }
-    end
-
-    def mirror_x?
-      %i[x both].include? mirror
-    end
-
-    def mirror_y?
-      %i[y both].include? mirror
-    end
-
-    def mirror_both?
-      mirror == :both
-    end
-  end
-end
